fbref_parser/requirements.txt
pandas
pyarrow
snakebite-py3
requests
lxml
beautifulsoup4
hdfs
